+++ packages/slice-machine/src/features/customTypes/customTypesBuilder/ImportSlicesFromLibraryModal/ImportSlicesFromLibraryModal.tsx
@@ -1,71 +1,42 @@
 import {
-  Box,
   Dialog,
-  DialogActionButton,
-  DialogActions,
-  DialogCancelButton,
   DialogContent,
   DialogDescription,
   DialogHeader,
-  Tab,
 } from "@prismicio/editor-ui";
 import { SharedSlice } from "@prismicio/types-internal/lib/customtypes";
-import { ReactNode, useState } from "react";
+import { useState } from "react";
 
-import { ComponentUI } from "@/legacy/lib/models/common/ComponentUI";
+import { LibrarySlicesDialogContent } from "./LibrarySlicesDialogContent";
+import { LocalSlicesDialogContent } from "./LocalSlicesDialogContent";
 
-import { getSubmitButtonLabel } from "../shared/getSubmitButtonLabel";
-import { useExistingSlices } from "../shared/useExistingSlices";
-import { useImportSlicesFromGithub } from "./hooks/useImportSlicesFromGithub";
-import { LibrarySlicesTab } from "./LibrarySlicesTab";
-import { LocalSlicesTab } from "./LocalSlicesTab";
-import {
-  ReuseExistingSlicesProvider,
-  useReuseExistingSlicesContext,
-} from "./ReuseExistingSlicesContext";
-
-interface ImportSlicesFromLibraryModalContent {
+interface ImportSlicesFromLibraryModalProps {
   open: boolean;
   location: "custom_type" | "page_type";
   typeName: string;
-  availableSlices?: ReadonlyArray<ComponentUI>;
+  availableSlices: (SharedSlice & { thumbnailUrl?: string })[];
   onSuccess: (args: { slices: SharedSlice[]; library?: string }) => void;
   onClose: () => void;
 }
 
-function ImportSlicesFromLibraryModalContent(
-  props: ImportSlicesFromLibraryModalContent,
+export function ImportSlicesFromLibraryModal(
+  props: ImportSlicesFromLibraryModalProps,
 ) {
   const {
     open,
-    location,
-    typeName,
     availableSlices = [],
     onSuccess,
     onClose,
+    ...commonProps
   } = props;
   const [selectedTab, setSelectedTab] = useState("local");
 
-  const existingSlices = useExistingSlices({ open });
-  const { resetSlices } = useImportSlicesFromGithub();
-  const { submit, isSubmitting, totalSelected, reset } =
-    useReuseExistingSlicesContext();
-
   const onOpenChange = (open: boolean) => {
-    if (open || isSubmitting) return;
+    if (open) return;
     onClose();
-    reset();
-    resetSlices();
+    // reset();
+    // resetSlices();
     setSelectedTab("local");
-  };
-
-  const onSubmit = async () => {
-    try {
-      const result = await submit({ existingSlices, location, resetSlices });
-      onSuccess(result);
-    } catch (error) {
-      // Error is already handled in the submit function
-    }
   };
 
   return (
@@ -75,81 +46,22 @@ function ImportSlicesFromLibraryModalContent(
         <DialogDescription hidden>
           Select existing slices or import slices from a GitHub repository
         </DialogDescription>
-        <Box flexDirection="column" flexGrow={1} minHeight={0}>
-          <Box
-            justifyContent="space-between"
-            padding={16}
-            border={{ bottom: true }}
-          >
-            <Box gap={8}>
-              <Tab
-                selected={selectedTab === "local"}
-                onClick={() => setSelectedTab("local")}
-              >
-                Local Slices
-              </Tab>
-              <Tab
-                selected={selectedTab === "library"}
-                onClick={() => setSelectedTab("library")}
-              >
-                Library Slices
-              </Tab>
-            </Box>
-            {selectedTab === "library" && "<repo-select>"}
-          </Box>
-          <TabContent selected={selectedTab === "local"}>
-            <LocalSlicesTab availableSlices={availableSlices} />
-          </TabContent>
-          <TabContent selected={selectedTab === "library"}>
-            <LibrarySlicesTab />
-          </TabContent>
-        </Box>
-
-        <DialogActions>
-          <DialogCancelButton disabled={isSubmitting} size="medium" />
-          <DialogActionButton
-            disabled={totalSelected === 0}
-            loading={isSubmitting}
-            onClick={() => void onSubmit()}
-            size="medium"
-          >
-            {getSubmitButtonLabel(location, typeName)} ({totalSelected})
-          </DialogActionButton>
-        </DialogActions>
+        <LocalSlicesDialogContent
+          {...commonProps}
+          isSelected={selectedTab === "local"}
+          onSelectTab={setSelectedTab}
+          availableSlices={availableSlices}
+          onSuccess={onSuccess}
+          onClose={onClose}
+        />
+        <LibrarySlicesDialogContent
+          {...commonProps}
+          isSelected={selectedTab === "library"}
+          onSelectTab={setSelectedTab}
+          onSuccess={onSuccess}
+          onClose={onClose}
+        />
       </DialogContent>
     </Dialog>
-  );
-}
-
-interface TabContentProps {
-  children: ReactNode;
-  selected: boolean;
-}
-
-function TabContent(args: TabContentProps) {
-  const { children, selected } = args;
-
-  if (!selected) {
-    return (
-      <Box display="none" minHeight={0}>
-        {children}
-      </Box>
-    );
-  }
-
-  return (
-    <Box display="flex" flexDirection="column" flexGrow={1} minHeight={0}>
-      {children}
-    </Box>
-  );
-}
-
-export function ImportSlicesFromLibraryModal(
-  props: ImportSlicesFromLibraryModalContent,
-) {
-  return (
-    <ReuseExistingSlicesProvider>
-      <ImportSlicesFromLibraryModalContent {...props} />
-    </ReuseExistingSlicesProvider>
   );
 }
